+++ R/gg_boxplot.R
@@ -644,7 +644,7 @@ gg_boxplot_col <- function(data,
       values = pal,
       drop = FALSE,
       labels = col_labels,
-      na.value = "#7F7F7FFF"
+      na.value = pal_na()
     ) 
   
   if (mobile == FALSE){
@@ -1381,7 +1381,7 @@ gg_boxplot_col_facet <- function(data,
       values = pal,
       drop = FALSE,
       labels = col_labels,
-      na.value = "#7F7F7FFF"
+      na.value = pal_na()
     ) +
     labs(
       title = stringr::str_wrap(title, title_wrap),