+++ R/gg_point.R
@@ -568,7 +568,7 @@ gg_point_col <- function(data,
       values = pal,
       drop = FALSE,
       labels = col_labels,
-      na.value = "#7F7F7FFF"
+      na.value = pal_na()
     ) 
   
   if (mobile == FALSE) {
@@ -1217,7 +1217,7 @@ gg_point_col_facet <-
         values = pal,
         drop = FALSE,
         labels = col_labels,
-        na.value = "#7F7F7FFF"
+        na.value = pal_na()
       ) +
       labs(
         title = stringr::str_wrap(title, title_wrap),