+++ R/pal.R
@@ -29,3 +29,15 @@ pal_viridis_reorder <- function(col_n) {
   else if(col_n == 2) viridis::viridis(4)[c(2, 3)]
   else if(col_n >= 3) viridis::viridis(col_n)
 }
+
+#' NA palette.
+#' 
+#' @description A function to retreive a hex code for a colour to use for NA values.
+#' 
+#' @return A vector of one hex code.
+#' @export 
+#' @examples 
+#' scales::show_col(pal_viridis_reorder(9)) 
+pal_na <- function() {
+  "#7F7F7FFF"
+}